fix: stop merging old log results and remove stale logs completely
The original idea behind the log files was to check them into a given repo - however, that's not very practical and the idea was dropped since they are cheap to rebuild. Unfortunately, the merging logic was left in place and resulted in some funky behavior where the logs seemed a run behind the latest.

Additionally, "stale" logs were also being reported on - which means even if you ran against a subset of tests you got reports for any/all of them that had run before - which was not intended.

These changes effectively ignore previous results and remove stale logs ensuring we only see reports for tests we ran and we don't leave any stale junk behind between runs.

Fixes: #2